requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.3
gunicorn==20.1.0
idna==2.10
pycryptodome==3.10.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
